+++ src/vector.h
@@ -7,7 +7,7 @@ typedef void* vector;
 
 vector vector_new(unsigned int base_size, size_t initial_cap);
 
-void* vector_get(vector self, size_t index);
+void* vector_get(vector self, unsigned int index);
 int vector_append(vector self, void *elem);
 int vector_remove(vector self, unsigned int index);
 size_t vector_size(vector self);
